C_Graphics_Projects
This repository is for my software projects done in only C language. The
software uploaded here were created using Turbo C and are MS DOS
compatible. We worked with these when we first learned C. So there can
be a lot of mistakes.

Update: Tetris is another game developed using C in Turbo C. It quite
resembles the original game. It was done in quite a hurry so it may
contain bugs.